NXP-29920: Extract JVM version check tests
